% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrierOptionPricing.R
\name{monte_carlo_arithmetic_asian}
\alias{monte_carlo_arithmetic_asian}
\title{Monte Carlo Arithmetic Asian Option Pricing}
\usage{
monte_carlo_arithmetic_asian(
  n_sim,
  S0,
  K,
  r,
  sigma,
  T,
  barrier,
  is_call,
  is_down_and_out
)
}
\arguments{
\item{n_sim}{Number of simulations}

\item{S0}{Initial stock price}

\item{K}{Strike price}

\item{r}{Risk-free rate}

\item{sigma}{Volatility}

\item{T}{Time to maturity}

\item{barrier}{Barrier level}

\item{is_call}{TRUE for call option, FALSE for put option}

\item{is_down_and_out}{TRUE for down-and-out, FALSE for up-and-out}
}
\value{
Option price
}
\description{
Monte Carlo Arithmetic Asian Option Pricing
}
\examples{
monte_carlo_arithmetic_asian(10000, 115, 100, 0.05, 0.25, 5/12, 90, TRUE, TRUE)
}
